Add in a value_type that is a pair of references.
I kinda want to go a different direction that meshes better with my preferred designs, but relative to the STL it would involve too much novel invention. The pair of references satisfies most use cases, even if it does have all the problems of proxy iterators.

This does make flat_map mostly just the same as map, though, which means wording for flat_map is much closer to complete now.